Move firstLogoUrl to a leaf, breaking an import cycle
Deleting genWithRetry took with it the import on line 2 of images.ts,
and the dev server stopped booting: circular dependency, on modules the
change never touched.

The cycle was already there — images → blog-site → referrals → credits
→ scheduler → director → content-preview → images — and stood only
because of the order the modules happened to initialise in, which that
line-2 import pinned. images.ts reached into blog-site.ts, the whole
public blog (Marked, the admin client, referrals), for one pure
five-line helper that picks the first logo out of an array.

Cut where the shared piece has no dependencies: firstLogoUrl now lives
in brand-fields.ts, the leaf with zero imports that exists for pure
functions over brand fields. Nothing else changes.

The check that caught it was npm run dev on the real route: 7289 unit
tests were green and svelte-check had nothing to say. An
initialisation cycle only shows itself when something runs.
